New User Screen 95% Complete!
The new user screen is now complete, I've made alot of modifications to the layout of the code in this file but finally happy with where it is. It now logs when users create others as well as logging if it was successful or if it failed, that will help with any troubleshooting required.

TODO:

- Ensure that only users with admin permission can create admins.